Create font providers interface and freetype package
