+++ Include/Rocket/Core/FreeType/FontProvider.h
@@ -25,19 +25,25 @@
  *
  */
 
-#ifndef ROCKETCOREFONTDATABASE_H
-#define ROCKETCOREFONTDATABASE_H
+#ifndef ROCKETCOREFREETYPEFONTPROVIDER_H
+#define ROCKETCOREFREETYPEFONTPROVIDER_H
 
 #include <Rocket/Core/StringUtilities.h>
 #include <Rocket/Core/Header.h>
 #include <Rocket/Core/Font.h>
+#include <Rocket/Core/FontProvider.h>
 
 namespace Rocket {
 namespace Core {
 
+
 class FontEffect;
 class FontFaceHandle;
-class FontProvider;
+class PropertyDictionary;
+
+namespace FreeType {
+
+class FontFamily;
 
 /**
 	The font database contains all font families currently in use by Rocket.
@@ -45,11 +51,31 @@ class FontProvider;
 	@author Peter Curry
  */
 
-class ROCKETCORE_API FontDatabase
+class ROCKETCORE_API FontProvider : public Rocket::Core::FontProvider
 {
 public:
 	static bool Initialise();
 	static void Shutdown();
+
+	/// Adds a new font face to the database. The face's family, style and weight will be determined from the face itself.
+	/// @param[in] file_name The file to load the face from.
+	/// @return True if the face was loaded successfully, false otherwise.
+	static bool LoadFontFace(const String& file_name);
+	/// Adds a new font face to the database, ignoring any family, style and weight information stored in the face itself.
+	/// @param[in] file_name The file to load the face from.
+	/// @param[in] family The family to add the face to.
+	/// @param[in] style The style of the face (normal or italic).
+	/// @param[in] weight The weight of the face (normal or bold).
+	/// @return True if the face was loaded successfully, false otherwise.
+	static bool LoadFontFace(const String& file_name, const String& family, Font::Style style, Font::Weight weight);
+	/// Adds a new font face to the database, loading from memory.
+	/// @param[in] data The font data.
+	/// @param[in] data_length Length of the data.
+	/// @param[in] family The family to add the face to.
+	/// @param[in] style The style of the face (normal or italic).
+	/// @param[in] weight The weight of the face (normal or bold).
+	/// @return True if the face was loaded successfully, false otherwise.
+	static bool LoadFontFace(const byte* data, int data_length, const String& family, Font::Style style, Font::Weight weight);
 
 	/// Returns a handle to a font face that can be used to position and render text. This will return the closest match
 	/// it can find, but in the event a font family is requested that does not exist, NULL will be returned instead of a
@@ -60,33 +86,26 @@ public:
 	/// @param[in] weight The weight of the desired font handle.
 	/// @param[in] size The size of desired handle, in points.
 	/// @return A valid handle if a matching (or closely matching) font face was found, NULL otherwise.
-	static FontFaceHandle* GetFontFaceHandle(const String& family, const String& charset, Font::Style style, Font::Weight weight, int size);
-
-	/// Returns a font effect, either a newly-instanced effect from the factory or an identical
-	/// shared effect.
-	/// @param[in] name The name of the desired font effect type.
-	/// @param[in] properties The properties associated with the font effect.
-	/// @return The requested font effect, or NULL if the font effect could not be found or instanced.
-	static FontEffect* GetFontEffect(const String& name, const PropertyDictionary& properties);
-
-	/// Removes a font effect from the font database's cache.
-	/// @param[in] The effect to release.
-	static void ReleaseFontEffect(const FontEffect* effect);
-    
-    /// Add a font provider to the database
-    /// @param[in] The provider to add.
-    static void AddFontProvider( FontProvider * provider );
+	virtual FontFaceHandle* GetFontFaceHandle(const String& family, const String& charset, Font::Style style, Font::Weight weight, int size);
 
 private:
-	FontDatabase(void);
-	~FontDatabase(void);
+	FontProvider(void);
+	~FontProvider(void);
 
-    typedef std::vector< FontProvider *> FontProviderTable;
+	// Adds a loaded face to the appropriate font family.
+	bool AddFace(void* face, const String& family, Font::Style style, Font::Weight weight, bool release_stream);
+	// Loads a FreeType face.
+	void* LoadFace(const String& file_name);
+	// Loads a FreeType face from memory.
+	void* LoadFace(const byte* data, int data_length, const String& source, bool local_data);
 
-    static FontProviderTable font_provider_table;
-	static FontDatabase* instance;
+	typedef std::map< String, FontFamily*, StringUtilities::StringComparei > FontFamilyMap;
+	FontFamilyMap font_families;
+
+	static FontProvider* instance;
 };
 
+}
 }
 }
