+++ Source/Core/ElementStyle.cpp
@@ -39,7 +39,7 @@
 #include "ElementBorder.h"
 #include "ElementDecoration.h"
 #include "ElementDefinition.h"
-#include "FontFaceHandle.h"
+#include <Rocket/Core/FontFaceHandle.h>
 
 namespace Rocket {
 namespace Core {
