+++ Source/Core/FontFaceLayer.cpp
@@ -28,7 +28,7 @@
 #include "precompiled.h"
 #include "FontFaceLayer.h"
 #include <Rocket/Core/Core.h>
-#include "FontFaceHandle.h"
+#include <Rocket/Core/FontFaceHandle.h>
 
 namespace Rocket {
 namespace Core {
