+++ Source/Core/UnicodeRange.cpp
@@ -26,7 +26,7 @@
  */
 
 #include "precompiled.h"
-#include "UnicodeRange.h"
+#include <Rocket/Core/UnicodeRange.h>
 
 namespace Rocket {
 namespace Core {
